rename userservice to exagrad-student-service
