+++ src/main/resources/application.yaml
@@ -1,6 +1,6 @@
 spring:
   application:
-    name: ASE-UserService
+    name: ASE-Exagrad-Student-Service
   jpa:
     database-platform: org.hibernate.dialect.H2Dialect
     show-sql: false
